Phase 4: extract service layer from fat models & views
Introduce ctrack/services/ and move business logic out of models and views.
Views become thin (parse -> service -> serialize); models keep data plus
trivial delegating accessors.

- import_service: absorbs Account.load_transactions orchestration and
  RecurringPayment.add_bill_from_file (wraps transaction_import / pdf_item_reader).
- categorisation_service: wraps ctrack.categories usage behind
  suggest_categories / get_clf_model / load_categoriser.
- progress_service: extracts the ~300-line ProgressView computation
  (resolve_period, compute_progress, _spend_by_category[_group]).
- reporting_service: moves the pandas-heavy Account balance and
  RecurringPayment bill-series reporting off the models.
- Drop the now-unused `from ctrack import categories` and
  `from ctrack.transaction_import import ...` top-level imports from models.py
  (plus pytz/datetime/time/timedelta no longer referenced there).

Behaviour-preserving: 123 tests green, no new migrations, OpenAPI schema
identical to the pre-change baseline.
